pyproject.toml
[project]
name = "web_research_agent"
version = "0.1.0"
description = "Deterministic research agent built with LangGraph"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "langgraph>=0.6",
  "pydantic>=2.0",
  "pydantic[email]>=2.0",
  "pyyaml>=6.0",
  "jsonschema>=4.0",
  "langfuse>=3.1.0",
  "exa-py>=2.0.0",
  "openai>=2.6.1",
  "python-dotenv>=1.2.1",
  "asyncpg>=0.30.0",
  "psycopg2-binary>=2.9.11",
  "sqlalchemy[asyncio]>=2.0.44",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.27.0",
  "email-validator>=2.0.0",
]

[project.optional-dependencies]
exa = ["exa-py"]
sonar = ["openai"]
all = ["exa-py", "openai", "langfuse>=3.1.0"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "tools*", "strategies*", "renderers*", "config*", "api*"]
